% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria_for_correlation.R
\name{bacteria_for_correlation}
\alias{bacteria_for_correlation}
\title{Example Dataset for bacteria_for_correlation}
\format{
A CSV file with 126 rows and 21 columns:
\describe{
\item{sample_id}{Sample identifier.}
\item{B_longum}{Relative abundance of \emph{Bifidobacterium longum}.}
\item{E_coli}{Relative abundance of \emph{Escherichia coli}.}
\item{B_bifidum}{Relative abundance of \emph{Bifidobacterium bifidum}.}
\item{B_breve}{Relative abundance of \emph{Bifidobacterium breve}.}
\item{k_pneumoniae}{Relative abundance of \emph{Klebsiella pneumoniae}.}
\item{S_salivarius}{Relative abundance of \emph{Streptococcus salivarius}.}
\item{S_lactarius}{Relative abundance of \emph{Streptococcus lactarius}.}
\item{E_faecium}{Relative abundance of \emph{Enterococcus faecium}.}
\item{S_mitis}{Relative abundance of \emph{Streptococcus mitis}.}
\item{S_peroris}{Relative abundance of \emph{Streptococcus peroris}.}
\item{S_pneumoniae}{Relative abundance of \emph{Streptococcus pneumoniae}.}
\item{S_lutetiensis}{Relative abundance of \emph{Streptococcus lutetiensis}.}
\item{k_variicola}{Relative abundance of \emph{Klebsiella variicola}.}
\item{B_fragilis}{Relative abundance of \emph{Bacteroides fragilis}.}
\item{S_equinus}{Relative abundance of \emph{Streptococcus equinus}.}
\item{C_sp}{Relative abundance of \emph{Clostridium} species.}
\item{P_cumminsii}{Relative abundance of \emph{Peptoniphilus cumminsii}.}
\item{T_bernardiae}{Relative abundance of \emph{Trueperella bernardiae}.}
\item{C_simulans}{Relative abundance of \emph{Corynebacterium simulans}.}
\item{G_SGB61413}{Relative abundance of \emph{Granulicatella} sp. SGB61413.}
}
}
\source{
Simulated data generated by the author.
}
\description{
A small synthetic dataset included with \emph{AssumpSure} for demonstrating
correlation analysis of the compositional data in the Correlation tab.
}
\details{
The file is stored in \code{inst/extdata/bacteria_for_correlation.csv}.

All values are synthetic, created only for demonstration; no human data were used,
and no ethical approval or consent is required.
}
\examples{
read.csv(system.file("extdata", "bacteria_for_correlation.csv", package = "AssumpSure"))
}
