% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the AssumpSure Shiny App}
\usage{
launch_app()
}
\value{
Runs the Shiny app.
}
\description{
Launch the AssumpSure Shiny App
}
\examples{
if (interactive()) {
  launch_app()
}
}
